Use latest versions of everything

+++ build.sh
@@ -3,10 +3,10 @@
 # TODO: https://emscripten.org/docs/porting/Debugging.html
 set -xeu
 
-PHP_VERSION=8.2.0RC5
+PHP_VERSION=8.2.5
 PHP_PATH=php-$PHP_VERSION
-AST_PATH=ast-1.1.0
-PHAN_VERSION=5.4.1
+AST_PATH=ast-1.1.2
+PHAN_VERSION=5.5.1
 PHAN_PATH=phan-$PHAN_VERSION.phar
 
 if ! type emconfigure 2>/dev/null >/dev/null ; then
@@ -14,12 +14,13 @@ if ! type emconfigure 2>/dev/null >/dev/null ; then
     exit 1
 fi
 
-echo "Get PHP source"
+echo "Check PHP source"
 if [ ! -d $PHP_PATH ]; then
     if [ ! -e $PHP_PATH.tar.xz ]; then
-        #wget https://www.php.net/distributions/$PHP_PATH.tar.xz
-        wget https://downloads.php.net/~pierrick/$PHP_PATH.tar.xz
+        echo "Get PHP source"
+        wget https://www.php.net/distributions/$PHP_PATH.tar.xz
     fi
+    echo "Extract PHP source"
     tar xf $PHP_PATH.tar.xz
 fi
 
@@ -31,8 +32,11 @@ echo "Get Phan phar"
 if [ ! -e $PHAN_PATH ]; then
     wget https://github.com/phan/phan/releases/download/$PHAN_VERSION/phan.phar -O $PHAN_PATH
 fi
+
+echo "Check ast"
 if [ ! -d "$PHP_PATH/ext/ast"  ]; then
     if [ ! -f "$AST_PATH.tgz" ]; then
+        echo "Get ast"
         wget https://pecl.php.net/get/$AST_PATH.tgz -O $AST_PATH.tgz
     fi
     tar zxf $AST_PATH.tgz
@@ -54,6 +58,8 @@ cd $PHP_PATH
 # Configure this with a minimal set of extensions, statically compiling the third-party ast library.
 # Run buildconf so that ast will a valid configure option
 ./buildconf --force
+
+set +e
 emconfigure ./configure \
   --disable-all \
   --disable-cgi \
@@ -77,10 +83,20 @@ emconfigure ./configure \
   --disable-fiber-asm \
   --enable-tokenizer
 
+if [ $? -ne 0 ]; then
+    echo "emconfigure failed. Content of config.log:"
+    cat config.log
+    exit 1
+fi
+
+set -e
+
 echo "Build"
-# -j5 seems to work for parallel builds
+
 emmake make clean
+# Debug failures with export EMCC_DEBUG=1; emmake make -j5 VERBOSE=1
 emmake make -j5
+
 rm -rf out
 mkdir -p out
 emcc $CFLAGS -I . -I Zend -I main -I TSRM/ ../pib_eval.c -c -o pib_eval.o
@@ -90,7 +106,7 @@ emcc $CFLAGS \
   --llvm-lto 2 \
   -s ENVIRONMENT=web \
   -s EXPORTED_FUNCTIONS='["_pib_eval", "_php_embed_init", "_zend_eval_string", "_php_embed_shutdown"]' \
-  -s EXTRA_EXPORTED_RUNTIME_METHODS='["ccall"]' \
+  -s EXPORTED_RUNTIME_METHODS='["ccall"]' \
   -s MODULARIZE=1 \
   -s EXPORT_NAME="'PHP'" \
   -s TOTAL_MEMORY=134217728 \
